+++ src/main/java/com/example/financeapp/modules/orderbook/OrderServiceImpl.java
@@ -192,7 +192,7 @@ public class OrderServiceImpl implements OrderService{
 
 
         Depot sellDepot = depotRepository.getById(sellOrder.getDepotId());
-        sellDepot.setTotalCash(sellDepot.getTotalCash()+(orderVolume*refP));
+        sellDepot.setTotalCash((int) (sellDepot.getTotalCash()+(orderVolume*refP)));
 
         Share shareToSell = shareRepository.getShareById(sellOrder.getShareId());
         DepotShareAmount sellDepotShareAmount = depotShareAmountRepository.findDepotShareAmountByDepotAndShare(sellDepot, shareToSell);
@@ -209,7 +209,7 @@ public class OrderServiceImpl implements OrderService{
 
 
         Depot buyDepot = depotRepository.getById(buyOrder.getDepotId());
-        buyDepot.setTotalCash(buyDepot.getTotalCash()-(orderVolume*refP));
+        buyDepot.setTotalCash((int) (buyDepot.getTotalCash()-(orderVolume*refP)));
         if(buyDepot.getTotalCash()<0){
             throw  new Exception("nicht genug geld um die Aktien zu kaufen");
         }
